Simplify testing
