+++ .travis.yml
@@ -6,15 +6,12 @@ os:
 #  - osx
 
 julia:
-  - 0.7
+  - 1.2
+  - 1.3
   - nightly
 
 notifications:
   email: false
-
-# https://github.com/travis-ci/travis-ci/issues/4942 workaround
-git:
-  depth: 99999999
 
 matrix:
   fast_finish: true
