+++ test/runtests.jl
@@ -1,11 +1,21 @@
-# Make sure the module source will be in the path whether this is run from
-# the julia REPL after installation or from a Makefile without being
-# installed.
-push!(LOAD_PATH, joinpath(dirname(@__FILE__), "..", "src"));
-# Then add the testing path so we can initialize the testing module
-push!(LOAD_PATH, joinpath(dirname(@__FILE__)));
+using Test
+const NumTypes = (Float32, Float64, BigFloat)
 
-import CMBTests
-CMBTests.loadall!();
-CMBTests.runtests();
+const TESTLIST = Dict(
+    "sphere" => "Sphere",
+    "legendre" => "Legendre",
+    "healpix" => "HEALPix",
+    "sparse" => "Sparse"
+   )
 
+@testset "CMB" begin
+    @testset "$desc" for (id,desc) in TESTLIST
+        modpath = joinpath(dirname(@__FILE__), "$(id).jl")
+        # Include the file and have it processed within this module
+        print("running $desc tests... ")
+        t0 = time_ns()
+        include(modpath)
+        t1 = time_ns()
+        println( (t1-t0)/1e9, " seconds")
+    end
+end
